requirements_backup.txt
pydantic-ai
pydantic-ai-slim[tavily]
streamlit
python-dotenv

